requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.5
fastapi==0.111.0
pydantic==2.7.1
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.21.0
uvicorn==0.29.0